+++ packages/client/src/components/business/ChatItem/ChatItem.tsx
@@ -8,10 +8,6 @@ import {
 import { ChatItemProps } from './types'
 import { useCustomIntl } from '@/hooks'
 
-// name = 'WakaWaka',
-//   image = 'https://image.shutterstock.com/image-photo/portrait-surprised-cat-scottish-straight-260nw-499196506.jpg',
-//   text = '06/09',
-
 export const ChatItem = ({ topic }: ChatItemProps) => {
   const { title, author } = topic
   return (
